+++ app/Jobs/InboxPipeline/InboxWorker.php
@@ -14,8 +14,9 @@ use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
-use Zttp\Zttp;
 use App\Jobs\DeletePipeline\DeleteRemoteProfilePipeline;
+use Illuminate\Support\Facades\Http;
+use Illuminate\Http\Client\ConnectionException;
 
 class InboxWorker implements ShouldQueue
 {
@@ -24,8 +25,9 @@ class InboxWorker implements ShouldQueue
     protected $headers;
     protected $payload;
 
-    public $timeout = 60;
+    public $timeout = 300;
     public $tries = 1;
+    public $maxExceptions = 1;
 
     /**
      * Create a new job instance.
@@ -47,10 +49,15 @@ class InboxWorker implements ShouldQueue
     {
         $profile = null;
         $headers = $this->headers;
+
+        if(empty($headers) || empty($this->payload) || !isset($headers['signature']) || !isset($headers['date'])) {
+            return;
+        }
+
         $payload = json_decode($this->payload, true, 8);
 
         if(isset($payload['id'])) {
-            $lockKey = hash('sha256', $payload['id']);
+            $lockKey = 'ap:icid:' . hash('sha256', $payload['id']);
             if(Cache::get($lockKey) !== null) {
                 // Job processed already
                 return 1;
@@ -58,18 +65,7 @@ class InboxWorker implements ShouldQueue
             Cache::put($lockKey, 1, 3600);
         }
 
-        if(!isset($headers['signature']) || !isset($headers['date'])) {
-            return;
-        }
-
-        if(empty($headers) || empty($payload)) {
-            return;
-        }
-
         if($this->verifySignature($headers, $payload) == true) {
-            (new Inbox($headers, $profile, $payload))->handle();
-            return;
-        } else if($this->blindKeyRotation($headers, $payload) == true) {
             (new Inbox($headers, $profile, $payload))->handle();
             return;
         } else {
@@ -84,18 +80,18 @@ class InboxWorker implements ShouldQueue
         $signature = is_array($headers['signature']) ? $headers['signature'][0] : $headers['signature'];
         $date = is_array($headers['date']) ? $headers['date'][0] : $headers['date'];
         if(!$signature) {
-            return;
+            return false;
         }
         if(!$date) {
-            return;
+            return false;
         }
         if(!now()->parse($date)->gt(now()->subDays(1)) || 
            !now()->parse($date)->lt(now()->addDays(1))
        ) {
-            return;
+            return false;
         }
         if(!isset($bodyDecoded['id'])) {
-        	return;
+        	return false;
         }
         $signatureData = HttpSignature::parseSignatureHeader($signature);
         $keyId = Helpers::validateUrl($signatureData['keyId']);
@@ -115,11 +111,11 @@ class InboxWorker implements ShouldQueue
                 }
             }
             if(parse_url($attr, PHP_URL_HOST) !== $keyDomain) {
-                return;
+                return false;
             }
         }
         if(!$keyDomain || !$idDomain || $keyDomain !== $idDomain) {
-            return;
+            return false;
         }
         $actor = Profile::whereKeyId($keyId)->first();
         if(!$actor) {
@@ -127,11 +123,11 @@ class InboxWorker implements ShouldQueue
             $actor = Helpers::profileFirstOrNew($actorUrl);
         }
         if(!$actor) {
-            return;
+            return false;
         }
         $pkey = openssl_pkey_get_public($actor->public_key);
         if(!$pkey) {
-            return 0;
+            return false;
         }
         $inboxPath = "/f/inbox";
         list($verified, $headers) = HttpSignature::verify($pkey, $signatureData, $headers, $inboxPath, $body);
@@ -166,10 +162,20 @@ class InboxWorker implements ShouldQueue
         if(Helpers::validateUrl($actor->remote_url) == false) {
             return;
         }
-        $res = Zttp::timeout(60)->withHeaders([
-          'Accept'     => 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
-          'User-Agent' => 'PixelfedBot v0.1 - https://pixelfed.org',
-        ])->get($actor->remote_url);
+
+        try {
+            $res = Http::timeout(20)->withHeaders([
+              'Accept'     => 'application/ld+json; profile="https://www.w3.org/ns/activitystreams"',
+              'User-Agent' => 'PixelfedBot v0.1 - https://pixelfed.org',
+            ])->get($actor->remote_url);
+        } catch (ConnectionException $e) {
+            return false;
+        }
+
+        if(!$res->ok()) {
+            return false;
+        }
+
         $res = json_decode($res->body(), true, 8);
         if(!$res || empty($res) || !isset($res['publicKey']) || !isset($res['publicKey']['id'])) {
         	return;