+++ app/Providers/AppServiceProvider.php
@@ -4,6 +4,7 @@ namespace App\Providers;
 
 use App\Observers\{
 	AvatarObserver,
+	FollowerObserver,
 	LikeObserver,
 	NotificationObserver,
 	ModLogObserver,
@@ -15,6 +16,7 @@ use App\Observers\{
 };
 use App\{
 	Avatar,
+	Follower,
 	Like,
 	Notification,
 	ModLog,
@@ -40,10 +42,14 @@ class AppServiceProvider extends ServiceProvider
 	 */
 	public function boot()
 	{
-		URL::forceScheme('https');
+		if(config('instance.force_https_urls', true)) {
+			URL::forceScheme('https');
+		}
+
 		Schema::defaultStringLength(191);
 		Paginator::useBootstrap();
 		Avatar::observe(AvatarObserver::class);
+		Follower::observe(FollowerObserver::class);
 		Like::observe(LikeObserver::class);
 		Notification::observe(NotificationObserver::class);
 		ModLog::observe(ModLogObserver::class);