+++ app/Services/ActivityPubFetchService.php
@@ -6,6 +6,7 @@ use Illuminate\Support\Facades\Http;
 use App\Profile;
 use App\Util\ActivityPub\Helpers;
 use App\Util\ActivityPub\HttpSignature;
+use Illuminate\Http\Client\ConnectionException;
 
 class ActivityPubFetchService
 {
@@ -16,12 +17,21 @@ class ActivityPubFetchService
 		}
 
 		$headers = HttpSignature::instanceActorSign($url, false);
-		$headers['Accept'] = 'application/activity+json, application/json';
+		$headers['Accept'] = 'application/activity+json, application/ld+json; profile="http://www.w3.org/ns/activitystreams"';
 		$headers['User-Agent'] = '(Pixelfed/'.config('pixelfed.version').'; +'.config('app.url').')';
 
-		return Http::withHeaders($headers)
-			->timeout(30)
-			->get($url)
-			->body();
+		try {
+			$res = Http::withHeaders($headers)
+				->timeout(10)
+				->get($url);
+		} catch (ConnectionException $e) {
+			return;
+		} catch (Exception $e) {
+			return;
+		}
+		if(!$res->ok()) {
+			return;
+		}
+		return $res->body();
 	}
 }