+++ app/Util/ActivityPub/Helpers.php
@@ -248,10 +248,13 @@ class Helpers {
 
 		$hash = hash('sha256', $url);
 		$key = "helpers:url:fetcher:sha256-{$hash}";
-		$ttl = now()->addMinutes(5);
+		$ttl = now()->addMinutes(15);
 
 		return Cache::remember($key, $ttl, function() use($url) {
 			$res = ActivityPubFetchService::get($url);
+			if(!$res || empty($res)) {
+				return false;
+			}
 			$res = json_decode($res, true, 8);
 			if(json_last_error() == JSON_ERROR_NONE) {
 				return $res;
@@ -496,7 +499,8 @@ class Helpers {
 				$status->in_reply_to_id === null &&
 				$status->reblog_of_id === null &&
 				in_array($status->type, ['photo', 'photo:album', 'video', 'video:album', 'photo:video:album']) &&
-				$status->created_at->gt(now()->subHours(config('instance.timeline.network.max_hours_old')))
+				$status->created_at->gt(now()->subHours(config('instance.timeline.network.max_hours_old'))) &&
+				(config('instance.hide_nsfw_on_public_feeds') == true ? $status->is_nsfw == false : true)
 			) {
 				NetworkTimelineService::add($status->id);
 			}
@@ -544,7 +548,7 @@ class Helpers {
 	public static function getScope($activity, $url)
 	{
 		$id = isset($activity['id']) ? self::pluckval($activity['id']) : self::pluckval($url);
-		$url = isset($activity['url']) ? self::pluckval($activity['url']) : $id;
+		$url = isset($activity['url']) ? self::pluckval($activity['url']) : self::pluckval($id);
 		$urlDomain = parse_url($url, PHP_URL_HOST);
 		$scope = 'private';