+++ config/instance.php
@@ -1,6 +1,7 @@
 <?php
 
 return [
+	'force_https_urls' => env('FORCE_HTTPS_URLS', true),
 
 	'description' => env('INSTANCE_DESCRIPTION', 'Pixelfed - Photo sharing for everyone'),
 
@@ -90,5 +91,11 @@ return [
 	'embed' => [
 		'profile' => env('INSTANCE_PROFILE_EMBEDS', true),
 		'post' => env('INSTANCE_POST_EMBEDS', true),
+	],
+
+	'hide_nsfw_on_public_feeds' => env('PF_HIDE_NSFW_ON_PUBLIC_FEEDS', false),
+
+	'avatar' => [
+		'local_to_cloud' => env('PF_LOCAL_AVATAR_TO_CLOUD', false)
 	],
 ];